removed matplotlib dependency for application
