+++ tools/config/config.py
@@ -20,12 +20,12 @@ use_mapper_selection = use_mapper_selection.lower()
 max_speed = 4 * 7.5  # set around 5-40 (normal-expert++)
 add_beat_intensity = 105  # try to match bps by x% [80, 120]
 cdf = 1.2  # cut director factor (to calculate speed, [0.5, 1.5])
-cdf_lr = 1.1  # speed addition factor for left right movement
+cdf_lr = 1.15  # speed addition factor for left right movement
 expert_fact = 0.63  # expert plus to expert factor [0.6, 0.7]
 create_expert_flag = True  # create second expert map
 thresh_beat = 0.42  # minimum beat response required to trigger generator [0.3, 0.6]
 thresh_pitch = 0.90  # minimum beat for pitch check (0.8,low-1.5,high)
-threshold_end = 1.0  # factor for start and end threshold [1.0, 1.2]
+threshold_end = 0.95  # factor for start and end threshold [1.0, 1.2]
 factor_pitch_certainty = 0.5  # select emphasis on first (>1) or second pitch method
 factor_pitch_meanmax = 3    # select pitch certainty for mean (>=3) or max (<3)
 random_note_map_factor = 0.3  # stick note map to random song/center (set to 0 to disable)
@@ -55,7 +55,7 @@ obstacle_max_count = 2  # maximum appearance count for obstacles
 sporty_obstacles = False
 add_slider_flag = True  # add arcs between notes in free areas
 slider_time_gap = [0.5, 10.0]    # time gap in seconds
-slider_probability = 0.9    # [0.1-1.0] with 1 meaning all on
+slider_probability = 0.92    # [0.1-1.0] with 1 meaning all on
 slider_movement_minimum = 1.5   # minimum movement between notes [0-5]
 slider_radius_multiplier = 1.0  # [0.5-2.0]
 slider_turbo_start = True   # start slider towards first note
