+++ training/eval_autoenc_music.py
@@ -1,4 +1,5 @@
 from helpers import *
+from plot_model import run_plot_autoenc
 from tensorflow_models import *
 from preprocessing.music_processing import run_music_preprocessing
 from tools.config import config, paths
